docs and refactoring
