+++ src/snakemake/workflow.py
@@ -19,7 +19,7 @@ from pathlib import Path
 import tarfile
 import tempfile
 from typing import Callable, Dict, Iterable, List, Optional, Set, Union
-from snakemake.io.access_patterns import AccessPatternFactory
+from snakemake.io.flags.access_patterns import AccessPatternFactory
 from snakemake.common.workdir_handler import WorkdirHandler
 from snakemake.settings.types import (
     ConfigSettings,
